+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.11.0] / 2026-02-11
+### Updates
+- Remove `Nuke.Common.Tools.MSBuild` to use `Nuke.Common.Tools.DotNet` for build and rebuild. (Fix: #85)
+- Add `GetConfigurations` in `net10` to get all configurations in the solution/project.
+- Disable warning in `net8.0`.
+
 ## [1.10.0] / 2026-01-12 - 2026-01-21
 ### Features
 - Update Nuke to version `9.0.4`.
@@ -421,6 +427,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.11.0]: ../../compare/1.10.0...1.11.0
 [1.10.0]: ../../compare/1.9.4...1.10.0
 [1.9.4]: ../../compare/1.9.3...1.9.4
 [1.9.3]: ../../compare/1.9.2...1.9.3